+++ src/main/java/cf/uiyoung/board/dao/UserDao.java
@@ -4,7 +4,7 @@ import cf.uiyoung.board.dto.User;
 
 public interface UserDao {
     User addUser(String email, String name, String password);
-    void mappingUserRole(Integer userId);
+    void mappingUserRole(Long userId);
     User getUser(String email);
     Integer updateUser(User user);
     Integer deleteUser(Integer userId);
